# repository/dtos.go
package repository

// Article is used as DTO for repository
type Article struct {
	ID      string `json:"id"`
	Title   string `json:"title"`
	Desc    string `json:"desc"`
	Content string `json:"content"`
}